Move FilterInput under Form
